Herencia y polimorfismo en los documentos carnet
