+++ Documentos/CarnetBase.cs
@@ -1,0 +1,61 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Drawing.Printing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace DSOO_PI1_ComB_Grupo15_Paez_Fernandez.Documentos
+{
+    public partial class CarnetBase : Form
+    {
+        public CarnetBase()
+        {
+            InitializeComponent();
+        }
+
+        protected void Carnet_Load(object sender, EventArgs e)
+        {
+            CargarDatos();
+        }
+
+        protected virtual void CargarDatos()
+        {
+
+        }
+
+        private void btnImprimir_Click(object sender, EventArgs e)
+        {
+            btnImprimir.Visible = false;
+
+            PrintDocument pd = new PrintDocument();
+            pd.PrintPage += new PrintPageEventHandler(ImprimirForm1);
+            pd.Print();
+
+            MessageBox.Show("Operación existosa", "AVISO DEL SISTEMA", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            Principal principal = new Principal();
+            principal.Show();
+            this.Close();
+
+            btnImprimir.Visible = true;
+        }
+
+        private void ImprimirForm1(object o, PrintPageEventArgs e)
+        {
+            int x = SystemInformation.WorkingArea.X;
+            int y = SystemInformation.WorkingArea.Y;
+            int ancho = this.Width;
+            int alto = this.Height;
+            Rectangle bounds = new Rectangle(x, y, ancho, alto);
+            Bitmap img = new Bitmap(ancho, alto);
+            this.DrawToBitmap(img, bounds);
+            Point p = new Point(100, 100);
+            e.Graphics.DrawImage(img, p);
+        }
+    }
+}
+
